[Payment]: 결제 승인 서비스 구현

+++ payment/src/main/java/com/devticket/payment/payment/application/dto/PgPaymentConfirmResult.java
@@ -5,7 +5,7 @@ public record PgPaymentConfirmResult(
     String orderId,
     String method,
     String status,
-    Long totalAmount,
+    Integer totalAmount,
     String approvedAt
 ) {
 }
