+++ payment/src/main/java/com/devticket/payment/payment/presentation/dto/PaymentConfirmRequest.java
@@ -13,6 +13,6 @@ public record PaymentConfirmRequest(
 
     @NotNull(message = "amount는 필수입니다.")
     @Positive(message = "amount는 0보다 커야 합니다.")
-    Long amount
+    Integer amount
 ) {
 }
